+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/LegacySessionManager.kt
@@ -516,20 +516,6 @@ class LegacySessionManager(
         values.find { session -> session.id == id }
     }
 
-    /**
-     * Informs this [SessionManager] that the OS is in low memory condition so it
-     * can reduce its allocated objects.
-     */
-    fun onLowMemory() {
-        // Removing the all the thumbnails except for the selected session to
-        // reduce memory consumption.
-        sessions.forEach {
-            if (it != selectedSession) {
-                it.thumbnail = null
-            }
-        }
-    }
-
     companion object {
         const val NO_SELECTION = -1
     }